move test to its own code
